ntkm_project_backend/WorkProjectNew/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==5.1.1
django-cors-headers==4.4.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
dnspython==2.6.1
fastapi==0.114.0
grid==0.7.1
idna==3.8
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.1.1
oauthlib==3.2.2
optional-django==0.3.0
psycopg==3.2.1
pycparser==2.22
pydantic==2.9.0
pydantic_core==2.23.2
PyJWT==2.9.0
python3-openid==3.2.0
react==4.3.0
requests==2.32.3
requests-oauthlib==2.0.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
starlette==0.38.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
